Moved all classes into the right package
