+++ src/main/java/org/example/springboot_labb2/entity/Message.java
@@ -1,4 +1,4 @@
-package org.example.springboot_labb2;
+package org.example.springboot_labb2.entity;
 
 
 import jakarta.persistence.*;
