+++ src/main/java/org/example/springboot_labb2/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package org.example.springboot_labb2;
+package org.example.springboot_labb2.repository;
 
+import org.example.springboot_labb2.entity.User;
 import org.springframework.data.repository.ListCrudRepository;
 import java.util.List;
 import java.util.UUID;
